add edition hook
