+++ frontend/src/screens/Editions/EditionsScreen.tsx
@@ -1,68 +1,28 @@
-import {
-  EditionsQuery,
-  useEditionsQuery,
-} from "../../graphql/editions.graphql.types";
+import { EditionsQuery } from "../../graphql/editions.graphql.types";
 import { Styles } from "../../utils/Styles";
-import { useState } from "react";
 import { Dialog } from "@mui/material";
 import { CloseHeader } from "../../components/dialogs/CloseHeader";
-import {
-  AddEditionForm,
-  EditionFormValues,
-} from "../../components/Editions/AddEditionForm";
-import { useCreateEditionMutation } from "../../graphql/createEdition.graphql.types";
-import { useDeleteEditionMutation } from "../../graphql/deleteEdition.graphql.types";
+import { AddEditionForm } from "../../components/Editions/AddEditionForm";
 import { EditionsList } from "../../components/Editions/EditionsList/EditionsList";
+import { useEditionsScreen } from "../../hooks/Editions/useEditionScreen";
 
 export type Edition = EditionsQuery["edition"][number];
 
 export const EditionsScreen = () => {
-  const { data, loading, error, refetch } = useEditionsQuery();
-  const editions: Edition[] = data?.edition ?? [];
-
-  const [isOpen, setIsOpen] = useState(false);
-  const [createEdition] = useCreateEditionMutation();
-  const [createError, setCreateError] = useState<string>();
-
-  const [deleteEdition] = useDeleteEditionMutation();
+  const {
+    editions,
+    loading,
+    error,
+    handleCreateClick,
+    createError,
+    isOpen,
+    openDialog,
+    closeDialog,
+    handleDeleteClick,
+  } = useEditionsScreen();
 
   if (loading) return <div>loading...</div>;
   if (error) return <div>ERROR: {error?.message}</div>;
-
-  const closeDialog = () => {
-    setCreateError(undefined);
-    setIsOpen(false);
-  };
-
-  const handleCreateClick = async (values: EditionFormValues) => {
-    try {
-      await createEdition({
-        variables: {
-          editionName: values.name,
-          editionYear: values.year,
-        },
-      });
-
-      refetch();
-      closeDialog();
-    } catch (error: unknown) {
-      console.error(error);
-
-      setCreateError(
-        error instanceof Error ? error.message : "Unexpected error received.",
-      );
-    }
-  };
-
-  const handleDeleteClick = async (id: string) => {
-    try {
-      await deleteEdition({ variables: { editionId: parseInt(id) } });
-      refetch();
-    } catch (error) {
-      // TODO some kind of global error ?
-      console.error(error);
-    }
-  };
 
   return (
     <div style={styles.container}>
@@ -74,7 +34,7 @@ export const EditionsScreen = () => {
         />
       </Dialog>
 
-      <button onClick={() => setIsOpen(true)}>dodaj edycje</button>
+      <button onClick={openDialog}>utwórz edycję</button>
 
       <EditionsList editions={editions} handleDeleteClick={handleDeleteClick} />
     </div>
